+++ src/GitHub_Updater/API/Gist_API.php
@@ -415,6 +415,7 @@ class Gist_API extends API implements API_Interface {
 		$this->type->owner   = $headers['owner'];
 		$this->type->slug    = $headers['repo'];
 		$this->type->gist_id = $headers['repo'];
+		$this->type->branch  = 'master';
 
 		$response         = $this->api( '/gists/:gist_id' );
 		$remote->meta     = $this->parse_meta_response( $response );