+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * use _dynamic_ constant for GitHub Updater plugin directory based on namespace
 * update to use `Languages` header as base for language pack packages, this should allow for self-hosted git servers
+* add header `Primary Branch` for those devs looking to replace `master`
 
 #### 9.6.1 / 2020-06-11
 * exit early from `Gist_API::construct_download_link()` if meta not present