+++ CHANGES.md
@@ -4,6 +4,7 @@
 * add header `Primary Branch` for those devs looking to replace `master`
 * fix PHP error when installing Gist by setting default branch to `master`
 * automatically add git host icons to plugin/theme row meta
+* update composer dependencies
 
 #### 9.6.1 / 2020-06-11
 * exit early from `Gist_API::construct_download_link()` if meta not present