+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 9.7.0 / 2020-07-09
 * use _dynamic_ constant for GitHub Updater plugin directory based on namespace
 * update to use `Languages` header as base for language pack packages, this should allow for self-hosted git servers
 * add header `Primary Branch` for those devs looking to replace `master`